Fix handling of modified root files
Summary:
This fixes a bug introduced on Metro `main` by D43774356 (no OSS releases).

`TreeFS.addOrModifyFile` did not properly canonicalise the path of the file to be modified, such that a modification or addition of a root-level file would set an entry in the internal map at (eg) `./App.js` instead of `App.js`. This caused subsequent lookups to fail, or return the old `App.js` metadata.

This diff runs the (relative, real) path through `normalizePath` a second time to account for the case of a parent canonical path being `.`.

Changelog:
* **[Fix]** Changes to files at the project root not properly handled by the watcher

Reviewed By: hoxyq

Differential Revision: D44462515

fbshipit-source-id: 74caf804b8f715488dc31721183d6625d7093e60