Fix missing content-length header when fetching assets (#961)
Summary:
This PR relates to PR https://github.com/facebook/metro/issues/953 and adds the `Content-Length` property when loading assets.

While this isn't mandatory for the apps, it's extremely useful for some debugging possibilities. At Expo, we are working on debugging improvements through Chrome DevTools, and one of these improvements is adding support for the Network inspector.

On the native side, we are limited to what we can infer automatically. One of these limitations is intercepting response bodies. We can't intercept these response bodies if it's larger than 1mb, and because of that, we rely on the `Content-Length` header to communicate the size of the response.

Pull Request resolved: https://github.com/facebook/metro/pull/961

Test Plan:
<img width="1312" alt="image" src="https://user-images.githubusercontent.com/1203991/228896070-61c4c3a4-578f-472a-b6e8-cfe7b2b8e624.png">

I can provide a repository to try this debugger, if it's necessary. You can also test this change by manually loading an existing asset (e.g. `http://localhost/assets/icon.png?platform=android`) and check the response headers in the same debugger.

Reviewed By: robhogan

Differential Revision: D44544314

Pulled By: huntie

fbshipit-source-id: 72da30b0822077e1dc5556a809bed5a67359595e