Tests: silence expected `console.warn` in `require-test` output
Summary:
Metro's `require` polyfill is expected to print require cycle warnings, and when testing cycle scenarios this spams the Jest output (except in the one case where we explicitly mock `console.warn` to check the warning).

This diff lifts the `console.warn` spy up so as to silence it.

Changelog: Internal

Reviewed By: jacdebug

Differential Revision: D44505719

fbshipit-source-id: 9ca845aa218d49896dcc2ec4723b403513302d80