Deprecate asyncRequireModulePath
Summary:
Changelog:

* **[Deprecated]**: The `transformer.asyncRequireModulePath` config option is deprecated. Use [`__loadBundleAsync`](https://github.com/react-native-community/discussions-and-proposals/blob/main/proposals/0605-lazy-bundling.md#__loadbundleasync-in-metro) instead.

Deprecates `asyncRequireModulePath` - a docs-only change for now.

At the moment, there is one valid use case for setting this option, namely pointing to the correct version of `metro-runtime` when `node_modules` resolution might not be reliable. The default RN Metro config does this. This use case is now considered "internal" and not publicly supported - when we fully remove the option we will also solve for reliable resolution.

Reviewed By: huntie

Differential Revision: D45001323

fbshipit-source-id: 17bf5997fb62853b3a87eaf0f4996dd9cb739df4