Don't assert that `paths` object is present in asyncRequire
Summary:
Makes it possible to provide a `__loadBundleAsync` implementations in bundles that don't have `experimentalImportBundleSupport` (or the upcoming `lazy` flag) enabled, and thus don't serialise a `paths` object to pass to `asyncRequire` calls. React Native injects one on `main` since D43600052 so this is technically a fix - it prevents the non-lazy `import()` behaviour from breaking.

Changelog:
NOTE: Merge changelog entry with D43600050 and link to both commits.
* **[Feature]**: Support custom `__loadBundleAsync` implementations in the default `asyncRequire` function. See the [lazy bundling RFC](https://github.com/react-native-community/discussions-and-proposals/blob/main/proposals/0605-lazy-bundling.md) for more details.

Reviewed By: huntie

Differential Revision: D45161543

fbshipit-source-id: 8a9624b8ee0e083179ee03b5fa8f2760b6e79147