Add "exports" special case handling for @babel/runtime (#990)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/990

This is a workaround for https://github.com/facebook/metro/issues/984. It adds a specific exception in the resolver for `babel/runtime` to prevent asserting the `"import"` condition name on these modules (when using Package Exports). This is necessary so that the CJS version of these Babel helpers are resolved, which enable CJS/MJS interop for all other modules (given our current strategy of resolving both `"require"` and `"import"` in all other packages and using Babel-driven ESM compatibility).

This workaround is removable if/when any of:
- https://github.com/babel/babel/pull/15643 is merged and updated in React Native.
- We implement dynamic handling of `"require"` and `"import"` conditions via ESM support in a future version of Metro.

Changelog: **[Experimental]** Fix `babel/runtime` issue when using Package Exports

Reviewed By: motiz88

Differential Revision: D46107056

fbshipit-source-id: 06c18de031458d6016ec7cf4101518e8bc1a174a