Avoid bail out of change event if modified time is zero
Summary:
Some watchers may not have the ability to detect modified time. In which case, these watchers can just return a null modified time to avoid the bailout logic that checks for changes in modified time before proceeding with change processing.

This means we also need to change some of the flow types for file metadata, as well as remove a few assertions that previously assumed file metadata always had a numeric modifiedTime.

Reviewed By: motiz88

Differential Revision: D46455003

fbshipit-source-id: 5a6ce33a3c591e30794bef2adc1bc4432d3228d2