Fix EPERM errors reading transient/locked files on Windows without Watchman
Summary:
As reported in https://github.com/facebook/metro/issues/1001, `fs.lstat` operations on Windows can fail with `EPERM` in certain circumstances - apparently when new files are either still being written or are about to be deleted.

Currently, this bubbles up as a Watcher error, whereas it should be handled gracefully.

Reportedly the OS will follow up with another event on the file when it's ready to read, so we may safely ignore `EPERM` in `add`/`change` event processing.

NOTE: This raises the question of why we're watching build output/intermediate directories in the default React Native Metro config. We should almost certainly be ignoring them from Metro.

```
 - **[Changelog]** Fix EPERM file watching errors on Windows.
```

Differential Revision: D46803761

fbshipit-source-id: e9551392208240e91ab813487efb94849f7e41b8