Consolidate `matchFilesWithContext` / `matchFiles` APIs
Summary:
We currently have `matchFiles` and `matchFilesWithContext` APIs within `FileSystem` (`TreeFS`) that operate in subtly different and surprising ways - eg whether posix and/or absolute paths are used as a basis for comparison, whether symlinks are followed, etc.

These differences are required at the call sites (`require.context` vs Jest tests matching patterns) but aren't clear on the interface.

This consolidates the two into a `matchFiles` API with a set of options allowing it to serve both use cases, reducing redundancy in the implementation and increasing coverage.  We also remove the concept of `context`-as-in-`require.context` from `FileSystem` and push it up to `DependencyGraph`. These APIs are internal.

Changelog: Internal.

Reviewed By: motiz88

Differential Revision: D46802785

fbshipit-source-id: c37ad2190d4cfa4a0d255ccdc3655fbbf09aa22c