Remove broken auto-conversion of resolver.blockList patterns
Summary:
D22598095 introduced a bug that partially broke `blockList` matching on Windows by introducing unescaped `\` characters into user-provided regexes. Since this logic *never* had the intended effect of normalising `/` characters in ignore patterns to `\\`, we remove it here.

Changelog: **[Fix]**: Fixed a bug in the handling of path separators in `blockList` on Windows.

Reviewed By: rozele

Differential Revision: D47060187

fbshipit-source-id: a238dd35f6bdb5f0a3b94c44d9edab57b5bb9654