Remove unused `metro-react-native-babel-preset` dependency
Summary:
Metro has no runtime dependencies on `metro-react-native-babel-preset`.

It looks like this was added way back in v0.40.x when `reactNativeTransformer` was part of the `metro` package. It's now decoupled (used when the path to `metro-react-native-babel-transformer` is passed through config), and we now only reference it in tests.

It's also installed as a project dev dependency in CLI user projects (for Jest's benefit) so I don't believe there's any risk this is propping up a missing dependency elsehwere.

This unblocks a move to the RN repo. CC dmytrorykun.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D47091438

fbshipit-source-id: 45c712a4541d7894e5758271e89301e7e9584065