Remove some $FlowIssues around tuple arity
Summary:
Flow is now able to narrow unions of tuples based on arity checks (`.length`). This removes a few unnecessary `$FlowIssue`s.

(Note: This doesn't work through the extra indirection of assigning `.length` to `n`, so I had to remove `n`, and then refactored to a switch to reduce repetitive checks - it does still work with `if/else if/...` though.)

Changelog: Internal

Reviewed By: huntie

Differential Revision: D47122425

fbshipit-source-id: d0a4c442052dca7081edda70862cfb71343e29d0