Use Flow comment syntax in metro-react-native-babel-transformer
Summary:
Now that `metro-react-native-babel-transformer` doesn't depend on `metro-source-maps`, it is isolated from the rest of Metro's Flow code and may be loaded without transformation if we use Flow comments in this one file.

This reduces some complexity (the need for Babel registration) in React Native's Jest preprocessor, as well as easing the move of this package into `react-native`. In particular, when the path to this module is passed by configuration to be loaded by Metro, and then Metro is run within RN integration tests from RN source - we don't have any clean mechanism for transforming it first.

Fortunately, using Flow comment here is quite light touch and we're able to do so without disabling prettier.

Changelog: Internal

Reviewed By: dmytrorykun

Differential Revision: D47162340

fbshipit-source-id: 545bce930591b4a2abda09d7cdb7e1f8f7cbb68c