Optimise/simplify getPlatformExtension
Summary:
`getPlatformExtension` extracts the `"ios"` from `"path/MyComponent.ios.js"`, if `"ios"` is a known platform. In the context of this function, this is only relevant to Haste.

Currently, the effective set of platforms is the union of those passed from Metro (from `config.resolver.platforms`) and a hardcoded set.

Entries in the hardcoded set but not in `config.resolver.platforms` are mostly superfluous because Metro will never resolve to them (although they would *prevent* `MyComponent.extraplatform` from being added to the Haste map with a generic platform).

The notable exception is `native`, which in Metro is a special platform alias used in the absence of a specific platform match.

Here, we remove the hardcoded set, and pass `native` from Metro when creating MetroFileMap.

Changelog: Internal

Reviewed By: motiz88

Differential Revision: D47050529

fbshipit-source-id: e9b5495e87133033a1f5270e36809c99eb75bbf9