metro-file-map: Separate MockMap from ModuleMap
Summary:
Ultimately we'd like to extract the concept of "mocks" from Metro since they're strictly a Jest thing, but with an eye on re-use of `metro-file-map` for Jest and/or upstreaming there's value in keeping them around, and covered by tests, but out of the way. This is a step towards extracting mock handling into an injected plugin or external subscriber.

The immediate motivation for this is to separate mocks from Haste - the two are unrelated and have almost no implementation overlap - so that mocks aren't in the way of refactoring Haste.

Changelog: Internal

Reviewed By: motiz88

Differential Revision: D45033203

fbshipit-source-id: 403b901ba4b3d92d163d937d41af1acdebf0b64e