Refactor metro-file-map tests to minimally depend on cached `haste`
Summary:
This refactors most tests to use the returned `hasteMap` and public `HasteMap` API rather than the captured cache content for asserting Haste state in various conditions.

A couple of exceptions remain for tests explicitly verifying that `haste` is persisted, which we'll remove when that's explicitly changed in the next diff.

Changelog: Internal

Reviewed By: motiz88

Differential Revision: D47600055

fbshipit-source-id: e0c3f0eca21388126f3afddda7bf6872ff2d3612