Preserve original locs of inlined requires
Summary:
Changelog: [Internal]

After inlining a `require` call, Metro currently records its *inlined* call sites as `locs` on the dependency object. However, for all build-time diagnostic purposes, it's more useful to record the *original* call site pre-inlining.

```
// Source code
const Foo = require('Foo');
//          ^^^^^^^^^^^^^^
//          Original loc

Foo(1);
^^^
// Inlined loc 1

Foo(2);
^^^
// Inlined loc 2

      ↓ ↓ ↓ ↓ ↓ ↓

// Transformed code
require('Foo')(1);

require('Foo')(2);
```

(Note that in some cases the original `require()` may itself come from an `import` declaration or other code. In general, we're interested in the user-authored source of the dependency.)

For step-by-step debugging purposes ( = in the source map) we still want to map each inlined call to the expression that it replaced (`Foo` identifier reference) and not to the original `require('Foo')`. This means we can't just clobber the `loc` on the `Foo` identifier with that of the original call. Instead, we **store a custom property on the AST** during the inline-requires pass, then read it during `collectDependencies` in preference to the standard `loc` property.

Custom AST properties are not safe to use in general because they do not survive printing / minification (and further transformation by other plugins). However, in this case:

1. The inline-requires plugin runs near the end of the transform pipeline, yet before the AST gets printed (or passed to the minifier).
2. The information passed through the custom property is not load-bearing; we can always reasonably fall back to the standard `loc` if it gets lost (at the cost of producing slightly worse diagnostics).

Reviewed By: jacdebug

Differential Revision: D47450006

fbshipit-source-id: 7c62bcbdbd0d5006d0d4cd30f03f79ccf09f61da