Remove runInspectorProxy option
Summary:
In preparation to relocate `metro-inspector-proxy` into the React Native repo, this PR removes the `runInspectorProxy` option from Metro's server, as a config option, and as a dependency of the `metro` package. This is a breaking change, where we now intend for dev servers that wrap Metro to mount this WebSocket middleware externally.

The `runInspectorProxy` option is unused today by RN CLI, Expo CLI, `rnx-kit`.

Changelog: **[Breaking]** Remove `runInspectorProxy` option from server

Reviewed By: motiz88

Differential Revision: D48066221

fbshipit-source-id: ce12ddbc1420bab9ddf47b1232433c1bc671c590