Fix tests dependent on Promise polyfill
Summary:
## `worker-test`

This is a test of the `buck-worker-tool`. Tests make use of an `end` utility function that waits for a number of chunks to be written to the worker's output (a JSON stream) before fulfilling. This diff updates the number of chunks that should be awaited to the correct value - 2 in each of these cases.

The first, important one is the command result, eg:
```
,{"type":"result","id":4,"exit_code":0}'
```

And the second chunk is written by the JSON stream (`stream.emit('root_end');`) on end:
```
]
```

So `2` is correct here.

The test currently passes because of the use of `new Promise().then()` in the test's `end()` utility function, and the fact that `metro-memory-fs` uses `process.nextTick` to simulate async I/O. With the polyfilled Promise, the `then()` callback is executed after a `process.nextTick`-based flush, which allows anything already queued with `process.nextTick` to execute first, so even though we haven't explicitly waited long enough, the work we assert on is completed just in time.

With real Promises, `then()` would be executed immediately on the microtask queue, and the test will fail unless all operations are complete before `end` resolves.

## `HttpStore-test`

As written, this test clearly *shouldn't* work. `jest.resetAllMocks()` resets `require('https').request` to a stub, so that attempting to add listeners to its return value, as `HttpStore.get` does immediately, will result in a `TypeError`.

It happens to work because the Promise polyfill [silently ignores unhandled rejections](https://github.com/then/promise#unhandled-rejections). With rejection tracking enabled, we reveal the error that's fatal with native promises.

```
  console.warn
      TypeError: Cannot read properties of undefined (reading 'on')

      at node_modules/promise/lib/rejection-tracking.js:105:13
```

What was intended here is `jest.clearAllMocks()`, to reset call counts. That fixes the error and no rejection is tracked.

Changelog: Internal

Reviewed By: motiz88

Differential Revision: D48186578

fbshipit-source-id: 8b033a4e2759eb5b2d7f7071764de653cebdbfc3