Add resolver.enableGlobalPackages config option
Summary:
Changelog:

* **[Feature]**: Add option to configure Metro's support for global package resolution (aka "Haste packages").

Metro supports a simple form of "workspace"-like resolution that ignores symlinks and folder structure, instead simply collecting all `package.json` files in the project into a global map. This will largely be superseded by symlink support when that ships as stable, but in advanced use cases, it can be desirable to disable this feature *independently* of whether symlinks are enabled. This diff makes the feature configurable (while leaving it on by default).

NOTE: This feature is known internally as "Haste package resolution" but it doesn't make sense to tie the external naming to Haste (an overloaded and poorly-understood term), so I'm going with the "global packages" naming that's already used in the test suite.

Reviewed By: robhogan

Differential Revision: D48349039

fbshipit-source-id: 417929516bcd7888219c10de7414bd7806b2a363