Fix Jest open handle in DeltaBundler resolver test
Summary:
This test was leaving dangling open handles in Jest because the "resolver" (containing a whole instance of `DependencyGraph`, watcher and all) wasn't properly ended.

`resolver` is automatically cleaned up in `afterEach` with `resolver.end()`, but in this test we were successfully creating a resolver instance without assigning it to `resolver`, leaving it hanging.

Changelog: Internal

Reviewed By: huntie

Differential Revision: D48783493

fbshipit-source-id: 711a2bc5879f4eb093e8ddf2fbc71a2804c7d976