Enable symlinks (`resolver.unstable_enableSymlinks`) by default (#1068)
Summary:
Pull Request resolved: https://github.com/facebook/metro/pull/1068

Flip the default for `unstable_enableSymlinks` to true. The option was made available in React Native 0.72 and it's been trialled by the community to the extent that we can have good confidence in it (though there are still devx gaps to guide users towards a suitable configuration).

This is also an effective prerequisite for disabling Haste/global package resolution by default in OSS, which currently allows workspaces to mostly-work in the absence of symlink support.

Closes https://github.com/facebook/metro/issues/1

Changelog:
```
 - **[Feature]**: Enable resolution through symlinks (previously experimental `unstable_enableSymlinks`)
```

Reviewed By: motiz88

Differential Revision: D48777855

fbshipit-source-id: a11664bac127db931c03d0e8446b6fc66e3c8a2b