Fix - removing an async dependency may incorrectly free its target
Summary:
Metro's core `Graph` maintains a `CountingSet` of inverse "eager" dependencies for each module - zero inverse dependencies triggers a module's removal from the graph.

Currently, this is not decremented with on dependency removals consistently with the way it is incremented. In particular, removing an async dependency (in combination with `options.lazy`) removes an inverse dependency, even though the reverse does not add one.

This leads to incorrect deletion of module B from the graph when module A has both sync and async dependencies on B, and the async dependency is removed.

```
* **[Fix]**: Fix Fast Refresh edge case where removing an async dependency could incorrectly remove modules from the graph.
```

Reviewed By: motiz88

Differential Revision: D50309038

fbshipit-source-id: 4a44703ac31fc400736c5666a51ca26783a34d19